+++ src/mlp.c
@@ -10,16 +10,16 @@ function lin = {linear, linear_deriv, LINEAR};
 function rel = {leaky_relu, leaky_relu_deriv, RELU};
 function softm = {softmax, softmax_deriv, SOFTMAX};
 
-double ranged_rand(double min, double max) {
-    return ((double)rand() / RAND_MAX) * (max - min) + min;
+float ranged_rand(float min, float max) {
+    return ((float)rand() / RAND_MAX) * (max - min) + min;
 }
 
 void init_neuron(neuron* neuron, int n_parameters) {
-    // neuron->inputs = malloc(n_parameters * sizeof(double));
-    neuron->weights = malloc(n_parameters * sizeof(double));
+    // neuron->inputs = malloc(n_parameters * sizeof(float));
+    neuron->weights = malloc(n_parameters * sizeof(float));
     neuron->n_weights = n_parameters;
 
-    double limit = sqrt(2.0 / n_parameters);
+    float limit = sqrt(2.0 / n_parameters);
 
     // Random init value for neurons and bias
     for (int i = 0; i < n_parameters; i++) {
@@ -33,19 +33,19 @@ layer dense(int n_neurons, int n_parameters, function *activation_function) {
     layer.activation_function = activation_function;
     layer.n_neurons = n_neurons;
     layer.neurons = malloc(n_neurons * sizeof(neuron));
-    layer.outputs = malloc(n_neurons * sizeof(double));
-    layer.raw_outputs = malloc(n_neurons * sizeof(double));
-    layer.derivatives = malloc(n_neurons * sizeof(double));
+    layer.outputs = malloc(n_neurons * sizeof(float));
+    layer.raw_outputs = malloc(n_neurons * sizeof(float));
+    layer.derivatives = malloc(n_neurons * sizeof(float));
 
     for (int i = 0; i < n_neurons; i++) { init_neuron(&layer.neurons[i], n_parameters); }
     return layer;
 }
 
 // Passage en avant (Forward Pass)
-void forward(MLP *m, double* inputs, int n_inputs) {
+void forward(MLP *m, float* inputs, int n_inputs) {
     for (int i = 0; i < m->n_layers; i++) {
         layer* l = &m->layers[i];
-        double* layer_inputs;
+        float* layer_inputs;
 
         // For input layer, input is the actual input
         if (i == 0) layer_inputs = inputs;
@@ -61,12 +61,12 @@ void forward(MLP *m, double* inputs, int n_inputs) {
 }
 
 // Entraînement (Backpropagation)
-void train(MLP *m, double* raw_inputs, double* target, double lr) {
+void train(MLP *m, float* raw_inputs, float* target, float lr) {
     // Calculate output layer error
     layer* l = &m->layers[m->n_layers - 1];
     l->activation_function->df(l->raw_outputs, l->derivatives, l->n_neurons);
     for (int i = 0; i < l->n_neurons; i++) {
-        // double output = l->neurons[i].output;
+        // float output = l->neurons[i].output;
         neuron* n = &l->neurons[i];
 
         // Error = (target - output) * f'(output)
@@ -82,7 +82,7 @@ void train(MLP *m, double* raw_inputs, double* target, double lr) {
 
         for (int j = 0; j < curr_layer->n_neurons; j++) {
             neuron* n = &curr_layer->neurons[j];
-            double error = 0;
+            float error = 0;
 
             // Sum deltas of next layer weighted by next neuron weights
             for (int k = 0; k < next_layer->n_neurons; k++) {
@@ -98,7 +98,7 @@ void train(MLP *m, double* raw_inputs, double* target, double lr) {
     // Update weights for each layer
     for (int i = 0; i < m->n_layers; i++) {
         layer* l = &m->layers[i];
-        double* inputs_for_layer;
+        float* inputs_for_layer;
 
         // For input layer, input is the actual input
         if (i == 0) inputs_for_layer = raw_inputs;
@@ -138,7 +138,7 @@ void print_model(MLP* m) {
     printf("Total number of parameters: %d\n", get_num_parameters(m));
 }
 
-void print_list(double* list, int len) {
+void print_list(float* list, int len) {
     printf("[");
     for (int i = 0; i < len; i++) {
         if (i+1 >= len) printf("%.2f", list[i]);
@@ -147,7 +147,7 @@ void print_list(double* list, int len) {
     printf("]");
 }
 
-void print_output(MLP *m, double* input, int input_len, double *expected, int expected_len) {
+void print_output(MLP *m, float* input, int input_len, float *expected, int expected_len) {
     layer* output = &m->layers[m->n_layers - 1];
     printf("Inputs: ");
     print_list(input, input_len);
@@ -159,7 +159,7 @@ void print_output(MLP *m, double* input, int input_len, double *expected, int ex
     printf("\n");
 }
 
-void one_hot(double* output, int input, int n_classes) {
+void one_hot(float* output, int input, int n_classes) {
     for (int i = 0; i < n_classes; i++) output[i] = 0; // Set/Reset buffer
     output[input] = 1.0;
 }
@@ -205,8 +205,8 @@ void save_model(MLP* m, const char* path) {
         for (int j = 0; j < l->n_neurons; j++) {
             neuron* n = &l->neurons[j];
             // Write weights and bias
-            fwrite(n->weights, sizeof(double), n->n_weights, f);
-            fwrite(&n->bias, sizeof(double), 1, f);
+            fwrite(n->weights, sizeof(float), n->n_weights, f);
+            fwrite(&n->bias, sizeof(float), 1, f);
         }
         fwrite(&l->activation_function->function_name, sizeof(int), 1, f);
     }
@@ -235,13 +235,13 @@ void load_model(MLP* m, const char* path) {
 
         l->neurons = malloc(l->n_neurons * sizeof(neuron));
 
-        l->outputs = malloc(l->n_neurons * sizeof(double));
+        l->outputs = malloc(l->n_neurons * sizeof(float));
 
-        l->raw_outputs = malloc(l->n_neurons * sizeof(double));
+        l->raw_outputs = malloc(l->n_neurons * sizeof(float));
 
         // This is just used for training, so we don't need if we're only doing inference
         // If you wish to train a model you loaded just replace the NULL by the commented malloc
-        l->derivatives = NULL; // malloc(l->n_neurons * sizeof(double));
+        l->derivatives = NULL; // malloc(l->n_neurons * sizeof(float));
 
         // Read number of weights
         fread(&l->neurons[0].n_weights, sizeof(int), 1, f);
@@ -251,11 +251,11 @@ void load_model(MLP* m, const char* path) {
             neuron* n = &l->neurons[j];
 
             n->n_weights = l->neurons[0].n_weights; // Same number of weights for each neuron of a given layer
-            n->weights = malloc(sizeof(double) * n->n_weights);
+            n->weights = malloc(sizeof(float) * n->n_weights);
 
             // Read weights and bias
-            fread(n->weights, sizeof(double), n->n_weights, f);
-            fread(&n->bias, sizeof(double), 1, f);
+            fread(n->weights, sizeof(float), n->n_weights, f);
+            fread(&n->bias, sizeof(float), 1, f);
         }
 
         int fn_name_buf;
